Export more than one file type in a single decamp
Added a feature to use an array of file extensions
in the Option ImageSuffix to Export the created
image with its settings into two images at once.

This way the .png and .eps file for example can be
obtained using ImageSuffix -> {".png",".eps"}
instead of two runs each using one of them.